refactor(security): split self-contained concerns out of security.ts
Extract the command-exec, URL/SSRF, and HTML/IP/redirect helpers from the
1167-line utils/security.ts into sibling modules: securityExec
(execFileSafe), securityUrl (validateUrl/isHostnameAllowed/allowlist +
private-IP blocklist), and securityHtml (sanitizeHtml/getClientIp/
validateRedirectUrl).

security.ts keeps the tightly-coupled path/fs core (resolveSafePath and the
safe fs wrappers) and re-exports the moved functions, so all 88 consumers
and 25 test mocks of `../utils/security` keep working unchanged. Behavior
is identical, including the nosemgrep annotations; the file drops
1167 -> 651 lines.
