+++ backend/src/utils/security.ts
@@ -1,4 +1,3 @@
-import { execFile } from "child_process";
 import fs from "fs-extra";
 import type {
   Dirent,
@@ -11,6 +10,17 @@ import {
   IMAGES_DIR,
   VIDEOS_DIR,
 } from "../config/paths";
+
+// Web/exec security helpers live in sibling modules; re-exported below so
+// existing `../utils/security` imports keep resolving the full surface.
+export { execFileSafe } from "./securityExec";
+export {
+  buildAllowlistedHttpUrl,
+  isHostnameAllowed,
+  validateUrl,
+  validateUrlWithAllowlist,
+} from "./securityUrl";
+export { getClientIp, sanitizeHtml, validateRedirectUrl } from "./securityHtml";
 
 /**
  * Safely rebuild a path from validated components while preserving absolute roots
@@ -563,11 +573,6 @@ export function sanitizePathSegment(segment: string): string {
 type ReadStreamOptions = Parameters<typeof fs.createReadStream>[1];
 type WriteStreamOptions = Parameters<typeof fs.createWriteStream>[1];
 type MoveSyncOptions = Parameters<typeof fs.moveSync>[2];
-type ExecFileSafeOptions = {
-  cwd?: string;
-  env?: NodeJS.ProcessEnv;
-  timeout?: number;
-};
 
 function normalizeAllowedDirectories(
   allowedDirOrDirs: string | readonly string[],
@@ -601,7 +606,6 @@ export async function removeImagePath(filePath: string): Promise<void> {
 
 /**
  * Resolves a child path inside an allowed directory.
- * Accepts relative path fragments and ensures the final path remains inside allowedDir.
  */
 export function resolveSafeChildPath(
   allowedDir: string,
@@ -644,524 +648,4 @@ export function normalizeSafeAbsolutePath(filePath: string): string {
  */
 export function validateCloudThumbnailCachePath(filePath: string): string {
   return resolveSafePath(filePath, CLOUD_THUMBNAIL_CACHE_DIR);
-}
-
-/**
- * Safely execute a command with arguments
- * Prevents command injection by using execFile instead of exec
- */
-export function execFileSafe(
-  command: string,
-  args: string[],
-  options?: ExecFileSafeOptions,
-): Promise<{ stdout: string; stderr: string }> {
-  return new Promise((resolve, reject) => {
-    execFile(command, args, options, (error, stdout, stderr) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve({ stdout: stdout.toString(), stderr: stderr.toString() });
-      }
-    });
-  });
-}
-
-const ALLOWED_URL_PROTOCOLS = new Set(["http:", "https:"]);
-const BLOCKED_PRIVATE_HOSTNAMES = new Set([
-  "localhost",
-  "0.0.0.0",
-  "::1",
-  "[::1]",
-  "::",
-]);
-const BLOCKED_PRIVATE_HOST_PREFIXES = [
-  "10.", // private (10.0.0.0/8)
-  "127.", // loopback (127.0.0.0/8)
-  "169.254.", // link-local incl. cloud metadata 169.254.169.254 (169.254.0.0/16)
-  "192.168.", // private (192.168.0.0/16)
-  ...Array.from({ length: 16 }, (_, i) => `172.${i + 16}.`), // 172.16.0.0/12
-  ...Array.from({ length: 64 }, (_, i) => `100.${i + 64}.`), // CGNAT 100.64.0.0/10
-];
-
-function normalizeHostnameForBlocklist(hostname: string): string {
-  return hostname.toLowerCase().replace(/\.+$/, "");
-}
-
-/**
- * Normalize non-dotted IPv4 encodings (e.g. decimal "2130706433" or hex
- * "0x7f000001", both = 127.0.0.1) to dotted-quad so range checks below cannot
- * be bypassed by an alternate numeric representation. Returns null when the
- * hostname is not a single-integer IPv4 form.
- */
-function normalizeNumericIpv4(hostname: string): string | null {
-  let value: number | null = null;
-  if (/^\d+$/.test(hostname)) {
-    value = Number(hostname);
-  } else if (/^0x[0-9a-f]+$/i.test(hostname)) {
-    value = Number.parseInt(hostname.slice(2), 16);
-  }
-
-  if (
-    value === null ||
-    !Number.isInteger(value) ||
-    value < 0 ||
-    value > 0xffffffff
-  ) {
-    return null;
-  }
-
-  return [
-    (value >>> 24) & 0xff,
-    (value >>> 16) & 0xff,
-    (value >>> 8) & 0xff,
-    value & 0xff,
-  ].join(".");
-}
-
-/**
- * Decode the IPv4 address embedded in an IPv4-mapped/compatible IPv6 literal.
- * Handles both the dotted form (`::ffff:169.254.169.254`) and the hex form the
- * WHATWG URL parser canonicalizes it to (`::ffff:a9fe:a9fe`). Returns null when
- * no trailing IPv4 is present.
- */
-function decodeEmbeddedIpv4(host: string): string | null {
-  const dotted = host.match(/(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})$/);
-  if (dotted) {
-    return dotted[1];
-  }
-
-  const hexPair = host.match(/:([0-9a-f]{1,4}):([0-9a-f]{1,4})$/);
-  if (hexPair) {
-    const high = Number.parseInt(hexPair[1], 16);
-    const low = Number.parseInt(hexPair[2], 16);
-    return [
-      (high >> 8) & 0xff,
-      high & 0xff,
-      (low >> 8) & 0xff,
-      low & 0xff,
-    ].join(".");
-  }
-
-  return null;
-}
-
-/**
- * Detect private/internal IPv6 literals: loopback (::1), unspecified (::),
- * unique-local (fc00::/7), link-local (fe80::/10), and IPv4-mapped/embedded
- * addresses (e.g. ::ffff:169.254.169.254) that wrap a blocked IPv4.
- */
-function isBlockedIpv6Hostname(hostname: string): boolean {
-  let host = hostname;
-  if (host.startsWith("[") && host.endsWith("]")) {
-    host = host.slice(1, -1);
-  }
-  if (!host.includes(":")) {
-    return false;
-  }
-  if (host === "::1" || host === "::") {
-    return true;
-  }
-
-  const firstHextet = host.split(":")[0];
-  // fc00::/7 (unique-local) -> first hextet begins fc or fd
-  // fe80::/10 (link-local)  -> first hextet begins fe8/fe9/fea/feb
-  if (/^f[cd]/.test(firstHextet) || /^fe[89ab]/.test(firstHextet)) {
-    return true;
-  }
-
-  // IPv4-mapped (e.g. ::ffff:127.0.0.1) embeds a blocked IPv4 address.
-  if (host.startsWith("::ffff:") || host.startsWith("::")) {
-    const embeddedIpv4 = decodeEmbeddedIpv4(host);
-    if (embeddedIpv4) {
-      return isBlockedPrivateHostname(embeddedIpv4);
-    }
-  }
-
-  return false;
-}
-
-function isBlockedPrivateHostname(hostname: string): boolean {
-  const normalized = normalizeHostnameForBlocklist(hostname);
-
-  if (BLOCKED_PRIVATE_HOSTNAMES.has(normalized)) {
-    return true;
-  }
-
-  if (isBlockedIpv6Hostname(normalized)) {
-    return true;
-  }
-
-  // Resolve alternate IPv4 encodings (decimal/hex) before prefix matching.
-  const candidate = normalizeNumericIpv4(normalized) ?? normalized;
-  if (BLOCKED_PRIVATE_HOSTNAMES.has(candidate)) {
-    return true;
-  }
-
-  return BLOCKED_PRIVATE_HOST_PREFIXES.some((prefix) =>
-    candidate.startsWith(prefix),
-  );
-}
-
-/**
- * Validates a URL to prevent SSRF attacks
- * Only allows http/https protocols and validates the hostname
- */
-export function validateUrl(url: string): string {
-  try {
-    const urlObj = new URL(url);
-
-    // Only allow http and https protocols
-    if (!ALLOWED_URL_PROTOCOLS.has(urlObj.protocol)) {
-      throw new Error(
-        `Invalid protocol: ${urlObj.protocol}. Only http and https are allowed.`,
-      );
-    }
-
-    // Block private/internal IP addresses
-    const hostname = urlObj.hostname;
-    if (isBlockedPrivateHostname(hostname)) {
-      throw new Error(
-        `SSRF protection: Blocked access to private/internal IP: ${hostname}`,
-      );
-    }
-
-    return url;
-  } catch (error) {
-    if (error instanceof TypeError) {
-      throw new Error(`Invalid URL format: ${url}`);
-    }
-    throw error;
-  }
-}
-
-/**
- * Allowed hostnames are exact (e.g. "missav.com") or subdomains (e.g. "www.missav.com").
- * Comparison is case-insensitive; hostname is normalized to lowercase.
- */
-export function isHostnameAllowed(
-  hostname: string,
-  allowedHostnames: readonly string[],
-): boolean {
-  const normalized = hostname.toLowerCase();
-  for (const allowed of allowedHostnames) {
-    const allowedLower = allowed.toLowerCase();
-    if (
-      normalized === allowedLower ||
-      normalized.endsWith("." + allowedLower)
-    ) {
-      return true;
-    }
-  }
-  return false;
-}
-
-/**
- * Rejects path traversal: pathname must not contain ".." as a path segment.
- */
-function hasPathTraversal(pathname: string): boolean {
-  const segments = pathname.split("/").filter((s) => s.length > 0);
-  return segments.some((segment) => segment === "..");
-}
-
-/**
- * Validates a URL for outgoing requests with an allow-list of hostnames to prevent SSRF.
- * Use this when the request target must be restricted to specific domains (e.g. a downloader for one site).
- * - Enforces http/https and blocks private IPs (same as validateUrl).
- * - Restricts hostname to the allow-list (exact or subdomain match).
- * - Rejects path traversal ("..") in the pathname.
- */
-export function validateUrlWithAllowlist(
-  url: string,
-  allowedHostnames: readonly string[],
-): string {
-  validateUrl(url);
-  const urlObj = new URL(url);
-  const hostname = urlObj.hostname;
-
-  if (!isHostnameAllowed(hostname, allowedHostnames)) {
-    throw new Error(
-      `SSRF protection: Hostname ${hostname} is not in the allow-list.`,
-    );
-  }
-
-  if (hasPathTraversal(urlObj.pathname)) {
-    throw new Error(
-      `SSRF protection: Path traversal ("..") is not allowed in the URL path.`,
-    );
-  }
-
-  return url;
-}
-
-/**
- * Builds a normalized http/https URL string after allow-list validation.
- * Result intentionally excludes credentials and explicit ports.
- */
-export function buildAllowlistedHttpUrl(
-  url: string,
-  allowedHostnames: readonly string[],
-): string {
-  const validatedUrl = validateUrlWithAllowlist(url, allowedHostnames);
-  const parsedUrl = new URL(validatedUrl);
-
-  if (!isHostnameAllowed(parsedUrl.hostname, allowedHostnames)) {
-    throw new Error(
-      `SSRF protection: Hostname ${parsedUrl.hostname} is not in the allow-list.`,
-    );
-  }
-
-  if (parsedUrl.username || parsedUrl.password || parsedUrl.port) {
-    throw new Error(
-      "SSRF protection: URLs with credentials or explicit ports are not allowed.",
-    );
-  }
-
-  return `${parsedUrl.protocol}//${parsedUrl.hostname}${parsedUrl.pathname}${parsedUrl.search}`;
-}
-
-/**
- * Sanitizes a string for safe use in HTML
- * Prevents XSS attacks
- */
-export function sanitizeHtml(str: string): string {
-  const map: { [key: string]: string } = {
-    "&": "&amp;",
-    "<": "&lt;",
-    ">": "&gt;",
-    '"': "&quot;",
-    "'": "&#x27;",
-    "/": "&#x2F;",
-  };
-  return str.replace(/[&<>"'/]/g, (s) => map[s]);
-}
-
-/**
- * Validates an IP address (IPv4 or IPv6)
- * @param ip - The IP address to validate
- * @returns true if the IP is valid, false otherwise
- */
-function isValidIpAddress(ip: string): boolean {
-  if (!ip || typeof ip !== "string") {
-    return false;
-  }
-
-  // IPv4 regex: matches 0.0.0.0 to 255.255.255.255
-  const ipv4Regex =
-    /^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/;
-
-  // IPv6 regex: matches various IPv6 formats including compressed notation
-  const ipv6Regex =
-    /^(?:[0-9a-fA-F]{1,4}:){7}[0-9a-fA-F]{1,4}$|^::1$|^::$|^(?:[0-9a-fA-F]{1,4}:)*::(?:[0-9a-fA-F]{1,4}:)*[0-9a-fA-F]{1,4}$/;
-
-  return ipv4Regex.test(ip) || ipv6Regex.test(ip);
-}
-
-/**
- * Checks if an IP address is a private/internal IP
- * @param ip - The IP address to check
- * @returns true if the IP is private/internal
- */
-function isPrivateIp(ip: string): boolean {
-  if (!ip || typeof ip !== "string") {
-    return false;
-  }
-
-  const cleanIp = ip.replace(/^::ffff:/, "");
-
-  // Check for localhost
-  if (
-    cleanIp === "localhost" ||
-    cleanIp === "0.0.0.0" ||
-    cleanIp === "::1" ||
-    cleanIp === "[::1]"
-  ) {
-    return true;
-  }
-
-  // Check for 127.x.x.x (loopback)
-  if (cleanIp.startsWith("127.")) {
-    return true;
-  }
-
-  // Check for 192.168.x.x (private network)
-  if (cleanIp.startsWith("192.168.")) {
-    return true;
-  }
-
-  // Check for 10.x.x.x (private network)
-  if (cleanIp.startsWith("10.")) {
-    return true;
-  }
-
-  // Check for 172.16.x.x to 172.31.x.x (private network)
-  if (cleanIp.startsWith("172.")) {
-    const parts = cleanIp.split(".");
-    if (parts.length >= 2) {
-      const secondOctet = parseInt(parts[1], 10);
-      if (secondOctet >= 16 && secondOctet <= 31) {
-        return true;
-      }
-    }
-  }
-
-  return false;
-}
-
-/**
- * Safely extracts the client IP address from a request
- * Prevents X-Forwarded-For header spoofing by validating the IP address
- *
- * Security considerations:
- * - Validates all IP addresses before using them
- * - Prioritizes socket IP (cannot be spoofed) over X-Forwarded-For
- * - Validates X-Forwarded-For header format and IP addresses
- * - When behind a proxy, uses X-Forwarded-For but validates it
- *
- * @param req - Express request object
- * @returns The validated client IP address
- */
-export function getClientIp(req: any): string {
-  // Get socket IP (the actual TCP connection IP - cannot be spoofed)
-  const socketIp = req.socket?.remoteAddress;
-  let cleanSocketIp: string | null = null;
-  if (socketIp && isValidIpAddress(socketIp)) {
-    const cleaned = socketIp.replace(/^::ffff:/, "");
-    if (isValidIpAddress(cleaned)) {
-      cleanSocketIp = cleaned;
-    }
-  }
-
-  // Check if we're behind a proxy (trust proxy is enabled)
-  // When behind a proxy, the socket IP is the proxy's IP, not the client's IP
-  const trustProxy = req.app?.get("trust proxy");
-  const isBehindProxy = trustProxy !== undefined && trustProxy !== false;
-
-  // If behind a proxy, try to get the client IP from X-Forwarded-For
-  // But we must validate it to prevent spoofing
-  if (isBehindProxy) {
-    const forwardedFor = req.headers["x-forwarded-for"];
-    if (forwardedFor && typeof forwardedFor === "string") {
-      // X-Forwarded-For format: "client, proxy1, proxy2"
-      // When trust proxy is set to 1, Express uses the rightmost IP
-      // But we need to validate all IPs to prevent spoofing
-      const ips = forwardedFor
-        .split(",")
-        .map((ip) => ip.trim())
-        .filter((ip) => ip.length > 0);
-
-      // When trust proxy is 1, we trust only the first proxy
-      // So we should use the rightmost IP (original client) or the first IP after the proxy
-      // For simplicity and security, we'll validate and use the rightmost valid IP
-      for (let i = ips.length - 1; i >= 0; i--) {
-        const ip = ips[i].trim();
-        const cleanIp = ip.replace(/^::ffff:/, "");
-
-        if (isValidIpAddress(cleanIp)) {
-          // Security: Only trust X-Forwarded-For when we're actually behind a proxy
-          // If socket IP is public (not behind proxy), ignore X-Forwarded-For to prevent spoofing
-          if (cleanSocketIp && isPrivateIp(cleanSocketIp)) {
-            // We're behind a proxy (socket IP is private)
-            // Trust X-Forwarded-For as it comes from the trusted proxy
-            // Use it regardless of whether it's public or private (validated IP)
-            return cleanIp;
-          } else if (cleanSocketIp && !isPrivateIp(cleanSocketIp)) {
-            // Socket IP is public - we're NOT behind a proxy
-            // Ignore X-Forwarded-For to prevent spoofing attacks
-            // This prevents attackers from bypassing rate limiting by spoofing X-Forwarded-For
-            break; // Exit loop, will use socket IP below
-          } else if (!cleanSocketIp) {
-            // Socket IP is missing or invalid, but we're configured to trust proxy
-            // In this case, we can use X-Forwarded-For as fallback
-            // But only if it's a valid IP format
-            return cleanIp;
-          }
-        }
-      }
-    }
-
-    // Fall back to req.ip if available (Express sets this when trust proxy is enabled)
-    if (req.ip && isValidIpAddress(req.ip)) {
-      const cleanIp = req.ip.replace(/^::ffff:/, "");
-      if (isValidIpAddress(cleanIp)) {
-        return cleanIp;
-      }
-    }
-  }
-
-  // If not behind a proxy, or X-Forwarded-For is invalid/missing, use socket IP
-  if (cleanSocketIp) {
-    return cleanSocketIp;
-  }
-
-  // Last resort: use socket IP even if validation failed
-  if (socketIp) {
-    return socketIp.replace(/^::ffff:/, "");
-  }
-
-  // If all else fails, return a default (this should rarely happen)
-  // Using "unknown" ensures rate limiting still works (all unknown IPs share the same limit)
-  return "unknown";
-}
-
-/**
- * Validates a redirect URL against an allowlist to prevent open redirect vulnerabilities
- * @param url - The URL to validate
- * @param allowedOrigin - The allowed origin (e.g., "https://example.com")
- * @returns The validated URL
- * @throws Error if the URL is invalid or not in the allowlist
- */
-export function validateRedirectUrl(
-  url: string,
-  allowedOrigin: string,
-): string {
-  // Ensure URL is a string and not empty
-  if (typeof url !== "string" || url.trim().length === 0) {
-    throw new Error("Redirect URL must be a non-empty string");
-  }
-
-  // Reject protocol-relative URLs (e.g., "//evil.com")
-  if (url.startsWith("//")) {
-    throw new Error("Protocol-relative URLs are not allowed");
-  }
-
-  // Reject dangerous protocols
-  const dangerousProtocols = ["javascript:", "data:", "vbscript:", "file:"];
-  const lowerUrl = url.toLowerCase().trim();
-  for (const protocol of dangerousProtocols) {
-    if (lowerUrl.startsWith(protocol)) {
-      throw new Error(`Dangerous protocol detected: ${protocol}`);
-    }
-  }
-
-  // Parse and validate the URL
-  let urlObj: URL;
-  try {
-    urlObj = new URL(url);
-  } catch (error) {
-    throw new Error(`Invalid URL format: ${url}`);
-  }
-
-  // Only allow http and https protocols
-  if (urlObj.protocol !== "http:" && urlObj.protocol !== "https:") {
-    throw new Error(
-      `Invalid protocol: ${urlObj.protocol}. Only http and https are allowed.`,
-    );
-  }
-
-  // Parse the allowed origin to get its origin
-  let allowedOriginObj: URL;
-  try {
-    allowedOriginObj = new URL(allowedOrigin);
-  } catch (error) {
-    throw new Error(`Invalid allowed origin format: ${allowedOrigin}`);
-  }
-
-  // Validate that the URL's origin matches the allowed origin exactly
-  if (urlObj.origin !== allowedOriginObj.origin) {
-    throw new Error(
-      `Redirect URL origin mismatch: ${urlObj.origin} is not allowed. Only ${allowedOriginObj.origin} is permitted.`,
-    );
-  }
-
-  return url;
 }
